+++ Sources/RKAPIService/RKAPIServiceProtocol.swift
@@ -16,39 +16,40 @@ public protocol RKAPIServiceProtocol {
      Fetch items with HTTP Get method without any body parameter. Uses async/await concurrency of iOS 13
      
      - Parameters:
-        - urlLink: Receives an optional URL
+        - urlLink: Receives an `Optional<URL>`
      
-     - Throws: An URLError is thrown if urlLink is nil or not a valied URL or server does not provide any response. Also ``HTTPStatusCode`` Error (Custom error) can be thrown if server status code is anything but 200...299
+     - Throws: An `URLError` is thrown if urlLink is nil or not a valied URL or server does not provide any response. Also ``HTTPStatusCode`` Error (Custom error) can be thrown if server status code is anything but 200...299
      
-     - Returns: Returns a  ``NetworkResult<T>`` where T is raw Data
+     - Returns: Returns a  ``NetworkResult``
      */
     @available(macOS 10.15.0, *)
     @available(iOS 13.0, *)
     func fetchItems(urlLink: URL?) async throws -> NetworkResult<Data>
     
     /**
+     Fetch items with HTTP Get method.
+     
      Fetch items with HTTP Get method without any body parameter. Uses async/await concurrency of iOS 13
      
      - Parameters:
         - urlLink: Receives an `Optional<URL>`
-        - completion: An `@escaping` closure parameter which provides a ``Result<NetworkResult<Data>, Error>`` as return of closure
+        - completion: An `@escaping` closure parameter which provides a `Result<Success, Failure>` where `Success` is ``NetworkResult`` and `Failure` is `Error` as return of closure
      */
     @available(iOS 9.0, *)
     @available(macOS 10.10, *)
     func fetchItems(urlLink: URL?, _ completion: @escaping (Result<NetworkResult<Data>, Error>)-> Void )
     
     /**
-     Fetch items with HTTP method with body parameter. Uses asyn/await method of iOS 13
+     Fetch items with HTTP method.
+     
+     Fetch items with HTTP method with body parameter. Uses asyn/await method of iOS 13.
      
      - Parameters:
         - urlLink: Receives an optional URL
         - httpMethod: ``HTTPMethod`` enum value to send data with that specific method.
         - body: Optional raw Data for sending to remote server.
         - jsonData: Accepts a boolean value to determine if HTTP body is in JSON format
-     
-     - Throws: An URLError is thrown if urlLink is nil or not a valied URL or server does not provide any response. Also ``HTTPStatusCode`` Error (Custom error) can be thrown if server status code is anything but 200...299
-     
-     - Returns: Returns a  ``NetworkResult<T>`` where T is raw Data
+        - completion: An `@escaping` closure parameter which provides a `Result<Success, Failure>` where `Success` is ``NetworkResult`` and `Failure` is `Error` as return of closure
      */
     @available(macOS 10.15.0, *)
     @available(iOS 13.0, *)
@@ -62,7 +63,7 @@ public protocol RKAPIServiceProtocol {
         - httpMethod: ``HTTPMethod`` enum value to send data with that specific method.
         - body: Optional raw Data for sending to remote server.
         - jsonData: Accepts a boolean value to determine if HTTP body is in JSON format
-        - completion: An `@escaping` closure parameter which provides a ``Result<NetworkResult<Data>, Error>`` as return of closure
+        - completion: An `@escaping` closure parameter which provides a `Result<NetworkResult<Data>, Error>` as return of closure
      */
     @available(iOS 9.0, *)
     @available(macOS 10.10, *)
@@ -76,7 +77,7 @@ public protocol RKAPIServiceProtocol {
      - Parameters:
         - urlLink: Receives an `Optional<URL>`
      
-     - Returns: Returns a  ``AnyPublisher<Success, Failure>`` where `Success` is ``NetworkResult`` `Failure` is ``Error``
+     - Returns: Returns a  `AnyPublisher<Success, Failure>` where `Success` is ``NetworkResult`` `Failure` is `Error`
      */
     @available(macOS 10.15.0, *)
     @available(iOS 13.0, *)
@@ -93,7 +94,7 @@ public protocol RKAPIServiceProtocol {
         - body: Optional raw Data for sending to remote server.
         - jsonData: Accepts a boolean value to determine if HTTP body is in JSON format
      
-     - Returns: Returns a  ``AnyPublisher<Success, Failure>`` where Success is ``NetworkResult`` Failure is ``Error``
+     - Returns: Returns a  `AnyPublisher<Success, Failure>` where Success is ``NetworkResult`` Failure is `Error`
      */
     @available(macOS 10.15.0, *)
     @available(iOS 13.0, *)